# 08-fullstack-portfolio-project/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Career Hub API",
    "version": "1.0.0",
    "description": "SI/SW 취업 준비용 풀스택 포트폴리오 Career Hub의 REST API 명세. https://editor.swagger.io 에 붙여 넣으면 시각적으로 확인할 수 있다.",
    "license": { "name": "MIT" }
  },
  "servers": [{ "url": "http://localhost:5100", "description": "로컬 개발" }],
  "tags": [
    { "name": "system", "description": "헬스 체크와 명세" },
    { "name": "auth", "description": "회원가입과 로그인" },
    { "name": "applications", "description": "지원 현황 CRUD" },
    { "name": "projects", "description": "포트폴리오 프로젝트 CRUD" },
    { "name": "dashboard", "description": "집계 지표" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "errors": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["message"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["token", "user"]
      },
      "Application": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "company": { "type": "string" },
          "role": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["준비중", "지원완료", "코딩테스트", "면접", "합격", "불합격"]
          },
          "priority": { "type": "string", "enum": ["낮음", "보통", "높음"] },
          "stack": { "type": "array", "items": { "type": "string" } },
          "dueDate": { "type": "string" },
          "contact": { "type": "string" },
          "memo": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "name": { "type": "string" },
          "summary": { "type": "string" },
          "status": { "type": "string", "enum": ["계획", "개발중", "완료"] },
          "stack": { "type": "array", "items": { "type": "string" } },
          "repoUrl": { "type": "string" },
          "deployUrl": { "type": "string" },
          "highlight": { "type": "string" }
        }
      },
      "Dashboard": {
        "type": "object",
        "properties": {
          "totalApplications": { "type": "integer" },
          "interviewCount": { "type": "integer" },
          "offerCount": { "type": "integer" },
          "upcomingCount": { "type": "integer" },
          "projectCount": { "type": "integer" },
          "completedProjectCount": { "type": "integer" },
          "statusCounts": { "type": "object", "additionalProperties": { "type": "integer" } }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "인증 실패",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Validation": {
        "description": "검증 실패",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "리소스를 찾을 수 없음",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": ["system"],
        "summary": "헬스 체크",
        "responses": {
          "200": {
            "description": "서버 정상",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "ok": { "type": "boolean" }, "service": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/openapi.json": {
      "get": {
        "tags": ["system"],
        "summary": "OpenAPI 명세 다운로드",
        "responses": { "200": { "description": "OpenAPI 3 문서" } }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "회원가입",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "가입 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "409": {
            "description": "이미 가입된 이메일",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "로그인",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": ["auth"],
        "summary": "내 정보",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "현재 사용자",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "user": { "$ref": "#/components/schemas/User" } }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/dashboard": {
      "get": {
        "tags": ["dashboard"],
        "summary": "대시보드 지표",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "집계 결과",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Dashboard" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/applications": {
      "get": {
        "tags": ["applications"],
        "summary": "지원 현황 목록",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Application" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["applications"],
        "summary": "지원 현황 추가",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Application" } }
          }
        },
        "responses": {
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Application" } }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/applications/{id}": {
      "patch": {
        "tags": ["applications"],
        "summary": "지원 현황 부분 수정",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Application" } }
          }
        },
        "responses": {
          "200": {
            "description": "수정됨",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Application" } }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["applications"],
        "summary": "지원 현황 삭제",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "삭제됨" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "프로젝트 목록",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "목록",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["projects"],
        "summary": "프로젝트 추가",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Project" } }
          }
        },
        "responses": {
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Project" } }
            }
          }
        }
      }
    },
    "/api/projects/{id}": {
      "patch": {
        "tags": ["projects"],
        "summary": "프로젝트 부분 수정",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Project" } }
          }
        },
        "responses": {
          "200": {
            "description": "수정됨",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Project" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "프로젝트 삭제",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "삭제됨" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  }
}